[js] replace web with web-app (+TS)
