--- web-app/.npmrc
@@ -1,1 +1,0 @@
-engine-strict=true
