+++ web/package.json
@@ -1,48 +1,45 @@
 {
-	"name": "web",
+	"name": "web-app",
 	"private": true,
 	"version": "0.0.1",
 	"type": "module",
-	"author": "Banshee, INC.",
-	"license": "Apache-2.0",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/banshee-data/velocity.report.git",
-		"directory": "web"
-	},
 	"scripts": {
-		"dev": "pnpm exec vite dev",
-		"build": "pnpm exec vite build",
-		"preview": "pnpm exec vite preview",
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
 		"prepare": "svelte-kit sync || echo ''",
-		"check": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json",
-		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json --watch",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"format": "prettier --write .",
 		"lint": "prettier --check . && eslint ."
-	},
-	"dependencies": {
-		"@layerstack/tailwind": "2.0.0-next.17",
-		"@layerstack/utils": "2.0.0-next.14",
-		"layerchart": "2.0.0-next.35",
-		"svelte-ux": "^2.0.0-next.17"
 	},
 	"devDependencies": {
 		"@eslint/compat": "^1.2.5",
 		"@eslint/js": "^9.18.0",
-		"@sveltejs/adapter-static": "^3.0.8",
+		"@layerstack/tailwind": "2.0.0-next.17",
+		"@sveltejs/adapter-auto": "^6.0.0",
 		"@sveltejs/kit": "^2.22.0",
 		"@sveltejs/vite-plugin-svelte": "^6.0.0",
-		"@tailwindcss/typography": "^0.5.15",
 		"@tailwindcss/vite": "^4.0.0",
 		"eslint": "^9.18.0",
 		"eslint-config-prettier": "^10.0.1",
 		"eslint-plugin-svelte": "^3.0.0",
 		"globals": "^16.0.0",
+		"layerchart": "2.0.0-next.35",
 		"prettier": "^3.4.2",
 		"prettier-plugin-svelte": "^3.3.3",
 		"prettier-plugin-tailwindcss": "^0.6.11",
 		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"svelte-ux": "2.0.0-next.18",
 		"tailwindcss": "^4.0.0",
+		"typescript": "^5.0.0",
+		"typescript-eslint": "^8.20.0",
 		"vite": "^7.0.4"
+	},
+	"pnpm": {
+		"onlyBuiltDependencies": [
+			"esbuild"
+		]
 	}
 }
