--- web/src/app.css
@@ -1,25 +1,0 @@
-@import 'tailwindcss';
-@plugin '@tailwindcss/typography';
-
-@import '@layerstack/tailwind/core.css';
-@import '@layerstack/tailwind/utils.css';
-@import '@layerstack/tailwind/themes/basic.css';
-
-@source '../../node_modules/svelte-ux/dist';
-@source '../../node_modules/layerchart/dist';
-
-/*
-You can customize your theme as you see fit using CSS variables.
-Dark mode is targeted using with .dark class on :root/<html>
-*/
-:root {
-	--color-primary: var(--color-purple-700);
-	--color-secondary: var(--color-orange-500);
-
-	&.dark {
-		--color-primary: var(--color-purple-400);
-		--color-surface-100: var(--color-zinc-900);
-		--color-surface-200: var(--color-zinc-950);
-		--color-surface-300: var(--color-black);
-	}
-}
