--- web/src/lib/index.js
@@ -1,1 +1,0 @@
-// place files you want to import through the `$lib` alias in this folder.
