+++ web/svelte.config.js
@@ -1,19 +1,17 @@
-import adapter from '@sveltejs/adapter-static';
+import adapter from '@sveltejs/adapter-auto';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	preprocess: [vitePreprocess({})],
+	// Consult https://svelte.dev/docs/kit/integrations
+	// for more information about preprocessors
+	preprocess: vitePreprocess(),
+
 	kit: {
-		// Using Static adapter for static site generation
-		// https://svelte.dev/docs/kit/adapter-static
-		adapter: adapter({
-			pages: 'build',
-			assets: 'build',
-			fallback: undefined, // @TODO: set 'index.html' for SPA fallback
-			precompress: false,
-			strict: true
-		})
+		// adapter-auto only supports some environments, see https://svelte.dev/docs/kit/adapter-auto for a list.
+		// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
+		// See https://svelte.dev/docs/kit/adapters for more information about adapters.
+		adapter: adapter()
 	}
 };
